Fix unittest of maepics (model)
